+++ CHANGELOG.md
@@ -1,10 +1,14 @@
 # Change Log
 
-## [Unreleased]
+## [0.8.4] - 2026-05-16
 
 - Added
   - Support inline class-attribute docstrings, add `DOC606` and `DOC607`, and
     introduce `--require-inline-class-var-docs` in native and flake8 modes
+  - A new config option `--skip-checking-private-functions` (shortform:
+    `-scpf`) to allow docstrings in private methods not to be checked
+- Full diff
+  - https://github.com/jsh9/pydoclint/compare/0.8.3...0.8.4
 
 ## [0.8.3] - 2025-11-26